[add]#03 コンポーネントごとにファイル作成
