+++ src/Board.jsx
@@ -1,0 +1,35 @@
+import { Square } from "./Square";
+
+
+export const Board = (props) => {
+  function renderSquare(i) {
+    return (
+      <Square
+        value={props.squares[i]}
+        onClick={() => {
+          props.onClick(i);
+        }}
+      />
+    );
+  }
+
+  return (
+    <div>
+      <div className='board-row'>
+        {renderSquare(0)}
+        {renderSquare(1)}
+        {renderSquare(2)}
+      </div>
+      <div className='board-row'>
+        {renderSquare(3)}
+        {renderSquare(4)}
+        {renderSquare(5)}
+      </div>
+      <div className='board-row'>
+        {renderSquare(6)}
+        {renderSquare(7)}
+        {renderSquare(8)}
+      </div>
+    </div>
+  );
+};
