+++ src/Square.jsx
@@ -1,0 +1,7 @@
+export const Square = (props) => {
+  return (
+    <button className='square' onClick={props.onClick}>
+      {props.value}
+    </button>
+  );
+};
