+++ src/index.js
@@ -1,125 +1,6 @@
-import React from 'react';
-import { useState } from 'react';
 import { createRoot } from 'react-dom/client';
+import { Game } from './Game';
 import './index.css';
-
-const Square = (props) => {
-  return (
-    <button className='square' onClick={props.onClick}>
-      {props.value}
-    </button>
-  );
-};
-
-const Board = (props) => {
-  function renderSquare(i) {
-    return (
-      <Square
-        value={props.squares[i]}
-        onClick={() => {
-          props.onClick(i);
-        }}
-      />
-    );
-  }
-
-  return (
-    <div>
-      <div className='board-row'>
-        {renderSquare(0)}
-        {renderSquare(1)}
-        {renderSquare(2)}
-      </div>
-      <div className='board-row'>
-        {renderSquare(3)}
-        {renderSquare(4)}
-        {renderSquare(5)}
-      </div>
-      <div className='board-row'>
-        {renderSquare(6)}
-        {renderSquare(7)}
-        {renderSquare(8)}
-      </div>
-    </div>
-  );
-};
-
-const Game = () => {
-  const [history, setHistory] = useState([{ squares: Array(9).fill(null) }]);
-  const [xIsNext, setXIsNext] = useState(true);
-  const [stepNumber, setStepNumber] = useState(0);
-
-  const current = history[stepNumber];
-
-  const winner = calcWinner(current.squares);
-
-  const moves = history.map((_, move) => {
-    const desc = move ? `Go to move #${move}` : 'Go to game start';
-    return (
-      <li key={move}>
-        <button onClick={() => jumpTo(move)}>{desc}</button>
-      </li>
-    );
-  });
-
-  function jumpTo(step) {
-    setStepNumber(step);
-    setXIsNext(step % 2 === 0);
-  }
-
-  function handleClick(i) {
-    const newSquares = [...current.squares];
-    if (newSquares[i] || winner) {
-      return;
-    }
-    newSquares[i] = xIsNext ? 'X' : 'O';
-    const pastHistory = history.slice(0, stepNumber + 1);
-    setHistory([...pastHistory, { squares: newSquares }]);
-    setXIsNext((prev) => !prev);
-    setStepNumber((prev) => prev + 1);
-  }
-
-  let status;
-  if (winner) {
-    status = `Winner: ${winner}`;
-  } else {
-    status = `Next player: ${xIsNext ? 'X' : 'O'}`;
-  }
-
-  return (
-    <div className='game'>
-      <div className='game-board'>
-        <Board squares={current.squares} onClick={(i) => handleClick(i)} />
-      </div>
-      <div className='game-info'>
-        <div>{status}</div>
-        <ol>{moves}</ol>
-      </div>
-    </div>
-  );
-};
-
-// ======================================== helper
-function calcWinner(squares) {
-  const lines = [
-    [0, 1, 2],
-    [3, 4, 5],
-    [6, 7, 8],
-    [0, 3, 6],
-    [1, 4, 7],
-    [2, 5, 8],
-    [0, 4, 8],
-    [2, 4, 6],
-  ];
-  for (const [a, b, c] of lines) {
-    if (squares[a] && squares[a] === squares[b] && squares[a] === squares[c]) {
-      return squares[a];
-    }
-  }
-  return null;
-}
-
-// ======================================== render
 
 const container = document.getElementById('root');
 const root = createRoot(container);
